Модуль специализации

+++ app/Modules/Specialization/Services/SpecializationService.php
@@ -3,14 +3,16 @@
 namespace App\Modules\Specialization\Services;
 
 use App\Modules\Specialization\Http\Requests\SpecializationRequest;
+use App\Modules\Specialization\Http\Requests\SpecializationUpdateRequest;
 use App\Modules\Specialization\Http\Resources\SpecializationResource;
 use App\Modules\Specialization\Interfaces\SpecializationRepositoryInterface;
 use App\Modules\Specialization\Interfaces\SpecializationServiceInterface;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Http\JsonResponse;
 
-class SpecializationService implements SpecializationServiceInterface
+readonly class SpecializationService implements SpecializationServiceInterface
 {
-    public function __construct(private readonly SpecializationRepositoryInterface $specializationRepository)
+    public function __construct(private SpecializationRepositoryInterface $specializationRepository)
     {
     }
 
@@ -30,24 +32,34 @@ class SpecializationService implements SpecializationServiceInterface
 
     public function getSpecializationById(int $id): SpecializationResource
     {
-        $specialization = $this->specializationRepository->getByIdWithServices($id);
+        $specialization = $this->specializationRepository->getById($id);
 
         return new SpecializationResource($specialization);
     }
 
-    public function updateSpecialization(SpecializationRequest $request, int $id): SpecializationResource
+    public function updateSpecialization(SpecializationUpdateRequest|SpecializationRequest $request, int $id): SpecializationResource
     {
-        $specialization = $this->specializationRepository->getByIdWithServices($id);
+        $specialization = $this->specializationRepository->getById($id);
         $updatedSpecialization = $this->specializationRepository->update($specialization, $request->validated());
 
         return new SpecializationResource($updatedSpecialization);
     }
 
-    public function deleteSpecialization(int $id): array
+    public function deleteSpecialization(int $id): JsonResponse
     {
-        $specialization = $this->specializationRepository->getByIdWithServices($id);
+        $specialization = $this->specializationRepository->getById($id);
+        $masterIds = $specialization->masters->pluck('id')->toArray();
+        if (count($masterIds) > 0) {
+            return response()->json([
+                'message' => "Нельзя удалить специализацию, так как с ней связаны мастера",
+                'code' => 409
+            ]);
+        }
         $this->specializationRepository->delete($specialization);
 
-        return ['message' => 'Специализация успешно удалена'];
+        return response()->json([
+            'message' => 'Специализация успешно удалена',
+            'code' => 200
+        ]);
     }
 }
